CrewAI/crewAI-landing_page_generator/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
crewai==0.11.0
dataclasses-json==0.6.4
datadispatch==1.0.0
deprecated==1.2.14
distro==1.9.0
docstring-parser==0.15
emoji==2.10.1
filetype==1.2.0
frozenlist==1.4.1
googleapis-common-protos==1.62.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-metadata==6.11.0
instructor==0.5.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.7
langchain-community==0.0.20
langchain-core==0.1.23
langchain-openai==0.0.2.post1
langdetect==1.0.9
langsmith==0.0.87
lxml==5.1.0
markdown-it-py==3.0.0
marshmallow==3.20.2
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.26.4
openai==1.12.0
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
packaging==23.2
protobuf==4.25.2
pydantic==2.6.1
pydantic-core==2.16.2
pygments==2.17.2
python-dotenv==1.0.0
python-iso639==2024.2.7
python-magic==0.4.27
pytils==0.4.1
pyyaml==6.0.1
rapidfuzz==3.6.1
regex==2023.12.25
requests==2.31.0
rich==13.7.0
setuptools==70.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlalchemy==2.0.27
tabulate==0.9.0
tasks==2.8.0
tenacity==8.2.3
tiktoken==0.5.2
tools==0.1.9
tqdm==4.66.2
typer==0.9.0
typing-extensions==4.9.0
typing-inspect==0.9.0
unstructured==0.10.25
urllib3==2.2.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
